Struct futures_iobuf::IoBuf

pub struct IoBuf {
    // some fields omitted
}

A reference counted buffer of bytes.

An IoBuf is a representation of a byte buffer where sub-slices of it can be handed out efficiently, each with a 'static lifetime which keeps the data alive. The buffer also supports mutation but may require bytes to be copied to complete the operation.

Methods

impl IoBuf

fn new() -> IoBuf

Creates a new IoBuf with no data and the default capacity.

fn with_capacity(cap: usize) -> IoBuf

Creates a new IoBuf with cap capacity.

fn len(&self) -> usize

Returns the number of bytes contained in this IoBuf.

fn as_slice(&self) -> &[u8]

Returns the inner contents of this IoBuf as a slice.

fn split_off(&mut self, at: usize) -> IoBuf

Splits the buffer into two at the given index.

Afterwards self contains elements [0, at), and the returned IoBuf contains elements [at, len).

This is an O(1) operation that just increases the reference count and sets a few indexes.

Panics

Panics if at > len

fn drain_to(&mut self, at: usize) -> IoBuf

Splits the buffer into two at the given index.

Afterwards self contains elements [at, len), and the returned IoBuf contains elements [0, at).

This is an O(1) operation that just increases the reference count and sets a few indexes.

Panics

Panics if at > len

fn get_mut(&mut self) -> IoBufMut

Returns a mutable reference to the underlying growable buffer of bytes.

If this IoBuf is the only instance pointing at the underlying buffer of bytes, a direct mutable reference will be returned. Otherwise the contents of this IoBuf will be reallocated in a fresh Vec<u8> allocation with the same capacity as this allocation, and that allocation will be returned.

This operation is not O(1) as it may clone the entire contents of this buffer.

The returned IoBufMut type implement Deref and DerefMut to Vec<u8> can the byte buffer can be manipulated using the standard Vec<u8> methods.